jwt_auth_requirements.txt
djoser
djangorestframework-simplejwt